requirements.txt
Flask==3.0.3
GitPython==3.1.43
pytest==8.1.2
